+++ geotrek/infrastructure/filters.py
@@ -1,10 +1,10 @@
 from django.contrib.contenttypes.models import ContentType
+from django_filters import CharFilter, MultipleChoiceFilter, ModelMultipleChoiceFilter
 from django.utils.translation import gettext_lazy as _
-from django_filters import CharFilter, MultipleChoiceFilter
 
 from geotrek.authent.filters import StructureRelatedFilterSet
 from geotrek.core.filters import ValidTopologyFilterSet, TopologyFilterTrail
-from .models import Infrastructure, INFRASTRUCTURE_TYPES
+from .models import Infrastructure, INFRASTRUCTURE_TYPES, InfrastructureMaintenanceDifficultyLevel, InfrastructureUsageDifficultyLevel
 from geotrek.maintenance.models import Intervention
 from geotrek.zoning.filters import ZoningFilterSet
 
@@ -18,6 +18,8 @@ class InfrastructureFilterSet(ValidTopologyFilterSet, ZoningFilterSet, Structure
     category = MultipleChoiceFilter(label=_("Category"), field_name='type__type',
                                     choices=INFRASTRUCTURE_TYPES)
     trail = TopologyFilterTrail(label=_('Trail'), required=False)
+    maintenance_difficulty = ModelMultipleChoiceFilter(queryset=InfrastructureMaintenanceDifficultyLevel.objects.all(), label=_("Maintenance difficulty"))
+    usage_difficulty = ModelMultipleChoiceFilter(queryset=InfrastructureUsageDifficultyLevel.objects.all(), label=_("Usage difficulty"))
 
     class Meta(StructureRelatedFilterSet.Meta):
         model = Infrastructure