+++ docs/changelog.rst
@@ -7,7 +7,14 @@ CHANGELOG
 
 **New features**
 
-- Add detailed jobs costs to interventions exports, with a new column for each job 
+- Add difficulty level fields (usage and maintenance) to infrastructure
+- Add 'active' field to job model, and hide inactive jobs in forms
+- Add detailed jobs costs to interventions exports, with a new column for each job
+
+**Bug fixes**
+
+* Fix length_kilometer attribute computation in treks
+* Fix date update format in lists
 
 
 2.62.0 (2021-07-06)
@@ -15,7 +22,7 @@ CHANGELOG
 
 **New features**
 
-- Add custom columns configuration to list views 
+- Add custom columns configuration to list views
 - Add custom columns configuration to list CSV exports
 - Add custom form fields configuration to creation views
 
@@ -63,7 +70,7 @@ CHANGELOG
 
 **Maintenance**
 
-- Update to django-mapentity 2.3.2
+- Update to paperclip 2.3.2
 
 
 2.59.0 (2021-06-07)